menu and permissions

+++ app/controllers/users_controller.rb
@@ -14,46 +14,58 @@ class UsersController < ApplicationController
 	end
 
 	def show
-		set_user
+		@user = User.find(params[:id])
 		render json: @user, methods: [:has_avatar, :avatar, :avatar_url_medium, :role]
 	end
 
 	def create
-		@user = User.new(user_params)
-		generated_password = Devise.friendly_token.first(8)
-		@user.password = generated_password
-		if @user.save!
-			#UserMailer.password_creation(@user, generated_password).deliver
-	    	render json: {success: true}
-	    else
-	      render json: @user.errors, status: :unprocessable_entity
-	    end  
-	end
-
-	def update
-		set_user
-		if user_params[:avatar] == "null"
-			@user.avatar = nil
-			params.delete :avatar
-		end
-		if @user.update(user_params)
-	      render json: {success: true}
-	    else
-	      render json: @user.errors, status: :unprocessable_entity
-	    end  
-	end
-
-	def destroy
-		set_user
-	    if @user.destroy
-	      render :json => {}
-	    else
-	      render :json =>  @user.errors, status: :unprocessable_entity
+		if current_user.confirm_permissions(["permission.manipulate_users"])
+			@user = User.new(user_params)
+			generated_password = Devise.friendly_token.first(8)
+			@user.password = generated_password
+			if @user.save!
+				#UserMailer.password_creation(@user, generated_password).deliver
+		    	render json: {success: true}
+		    else
+		      render json: @user.errors, status: :unprocessable_entity
+		    end
+		else
+	    	render json: {error: "Unauthorized access"}, status: 401
 	    end
 	end
 
+	def update
+		@user = User.find(params[:id])
+		if @user.id == current_user.id || current_user.confirm_permissions(["permission.manipulate_users"])
+			if user_params[:avatar] == "null"
+				@user.avatar = nil
+				params.delete :avatar
+			end
+			if @user.update(user_params)
+		      render json: {success: true}
+		    else
+		      render json: @user.errors, status: :unprocessable_entity
+		    end
+		else
+			render json: {error: "Unauthorized access"}, status: 401
+		end
+	end
+
+	def destroy
+		if current_user.confirm_permissions(["permission.manipulate_users"])
+			@user = User.find(params[:id])
+		    if @user.destroy
+		      render :json => {}
+		    else
+		      render :json =>  @user.errors, status: :unprocessable_entity
+		    end
+		else
+			render json: {error: "Unauthorized access"}, status: 401
+		end
+	end
+
 	def verify_permissions
-		set_user
+		@user = User.find(params[:id])
 		verified = @user.verify_permissions(params[:permissions])
 		render json: verified
 	end
@@ -62,10 +74,6 @@ class UsersController < ApplicationController
 
 		def user_params
 			params.permit(:id, :first_name, :last_name, :email, :birthdate, :avatar, :about, :role_id)
-		end
-
-		def set_user
-			@user = User.find(params[:id])
 		end
 
 end
